+++ ChangeLog.md
@@ -2,6 +2,14 @@
 
 ## Versions ##
 
+### Unreleased ###
+
+* Adds bulleted/numbered list support - Example39.ps1 to Example44.ps1 (#105)
+* Adds custom list number formats (Word and Text plugins only)
+  * Adds `NumberStyle` keyword
+* Fixes bug in Text table output breaking Word and Html output (#126)
+* Fixes bug in call stack enumeration with nested functions containing `Section -Orientation` definitions (#121)
+
 ### 0.10.0 ###
 
 * __BREAKING CHANGE__ - XML output to be deprecated in a future release (#102)